Implemented a large part of covering. Now, rules towards actions can be formed. Still have yet to create and unify those rules, but the framework is in place. Also created MultiMap, a class which behaves like Map<K, List<V>>, which I have been meaning to create for some time.

+++ MsPacMan/src/rlPacMan/PacManExperiment.java
@@ -65,8 +65,8 @@ public class PacManExperiment {
         }
         RLGlue.RL_agent_message("unfreeze learning");
 
-        mean = sum / (double)n;
-        variance = (sum_of_squares - (double)n * mean * mean) / ((double)n - 1.0f);
+        mean = sum / n;
+        variance = (sum_of_squares - n * mean * mean) / ((double)n - 1.0f);
         return new EvaluationPoint(mean, Math.sqrt(variance));
     }
     /*
